Some cleanup.  Minor refactorying
