+++ club-oauth2/src/main/java/smh/club/oauth2/config/mapper/ObjectMapperConfig.java
@@ -5,7 +5,6 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;
 import org.springframework.security.jackson2.SecurityJackson2Modules;
 import org.springframework.security.oauth2.server.authorization.jackson2.OAuth2AuthorizationServerJackson2Module;
-import smh.club.oauth2.domain.entities.UserDetailsEntity;
 
 @Configuration
 public class ObjectMapperConfig {
@@ -17,6 +16,6 @@ public class ObjectMapperConfig {
     var modules = SecurityJackson2Modules.getModules(classLoader);
     modules.add(new OAuth2AuthorizationServerJackson2Module());
 
-    return  new Jackson2ObjectMapperBuilder().modules(modules).mixIn(UserDetailsEntity.class, UserEntityMixin.class);
+    return  new Jackson2ObjectMapperBuilder().modules(modules);
   }
 }
