+++ club-oauth2/src/test/java/smh/club/oauth2/AuthorizationServerApplicationIntegrationTests.java
@@ -1,4 +1,4 @@
-package smh.club.oauth2.config;
+package smh.club.oauth2;
 
 import io.zonky.test.db.AutoConfigureEmbeddedDatabase;
 import java.io.IOException;
